Migrate to Cassandra (tests not working!)

+++ nodehistj-newest-nodelists/src/main/java/ru/oldzoomer/nodehistj_newest_nodelists/util/NodelistFillToDatabase.java
@@ -16,18 +16,15 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import ru.oldzoomer.minio.utils.MinioUtils;
 import ru.oldzoomer.nodehistj_newest_nodelists.entity.NodeEntry;
-import ru.oldzoomer.nodehistj_newest_nodelists.entity.NodelistEntry;
 import ru.oldzoomer.nodehistj_newest_nodelists.exception.NoNewObjects;
 import ru.oldzoomer.nodehistj_newest_nodelists.repo.NodeEntryRepository;
-import ru.oldzoomer.nodehistj_newest_nodelists.repo.NodelistEntryRepository;
 import ru.oldzoomer.nodelistj.Nodelist;
 import ru.oldzoomer.redis.utils.ClearRedisCache;
 
 /**
  * Component for processing and storing historical nodelists in the database.
  * Listens to Kafka topic for new nodelist files, downloads them from MinIO
- * storage,
- * parses and saves to database, then clears Redis cache.
+ * storage, parses and saves to database, then clears Redis cache.
  */
 @RequiredArgsConstructor
 @Component
@@ -35,7 +32,6 @@ import ru.oldzoomer.redis.utils.ClearRedisCache;
 public class NodelistFillToDatabase {
     private final MinioUtils minioUtils;
     private final NodeEntryRepository nodeEntryRepository;
-    private final NodelistEntryRepository nodelistEntryRepository;
     private final ClearRedisCache clearRedisCache;
 
     /** MinIO bucket name where nodelists are stored */
@@ -44,14 +40,15 @@ public class NodelistFillToDatabase {
 
     /**
      * Converts nodelist entry from common format to database entity
-     * 
-     * @param nodeListEntry    source nodelist entry from common library
-     * @param nodelistEntryNew parent nodelist entry entity
+     *
+     * @param nodeListEntry source nodelist entry from common library
+     * @param year          the year of the nodelist
+     * @param name          the name of the nodelist file
      * @return populated NodeEntry entity ready for saving
      */
     private static NodeEntry getNodeEntry(
             ru.oldzoomer.nodelistj.entries.NodelistEntry nodeListEntry,
-            ru.oldzoomer.nodehistj_newest_nodelists.entity.NodelistEntry nodelistEntryNew) {
+            Integer year, String name) {
         NodeEntry nodeEntryNew = new NodeEntry();
 
         nodeEntryNew.setZone(nodeListEntry.zone());
@@ -64,46 +61,53 @@ public class NodelistFillToDatabase {
         nodeEntryNew.setPhone(nodeListEntry.phone());
         nodeEntryNew.setSysOpName(nodeListEntry.sysOpName());
         nodeEntryNew.setFlags(Arrays.asList(nodeListEntry.flags()));
-        nodeEntryNew.setNodelistEntry(nodelistEntryNew);
+        nodeEntryNew.setNodelistYear(year);
+        nodeEntryNew.setNodelistName(name);
         return nodeEntryNew;
     }
 
     /**
      * Kafka listener method triggered when new nodelists are downloaded.
      * Processes each modified nodelist file from MinIO storage.
-     * 
+     *
      * @param modifiedObjects list of MinIO object paths that were modified
      */
+    @Transactional
     public void updateNodelist(List<String> modifiedObjects) {
         log.info("Update nodelists is started");
 
-        String modifiedObject = modifiedObjects.stream()
-                .filter(object -> object.matches(".*/\\d{4}/nodelist\\.\\d{3}"))
-                .sorted(Comparator.reverseOrder())
-                .findFirst()
-                .orElseThrow(NoNewObjects::new);
+        try {
+            String modifiedObject = modifiedObjects.stream()
+                    .filter(object -> object.matches(".*/\\d{4}/nodelist\\.\\d{3}"))
+                    .sorted(Comparator.reverseOrder())
+                    .findFirst()
+                    .orElseThrow(NoNewObjects::new);
 
-        log.info("New object: {}", modifiedObject);
+            log.info("New object: {}", modifiedObject);
 
-        try (InputStream inputStream = minioUtils.getObject(minioBucket, modifiedObject)) {
-            Matcher matcher = Pattern.compile(".*/(\\d{4})/(nodelist\\.\\d{3})").matcher(modifiedObject);
-            matcher.find(); // Initializes the matcher
-            Nodelist nodelist = new Nodelist(new ByteArrayInputStream(inputStream.readAllBytes()));
+            try (InputStream inputStream = minioUtils.getObject(minioBucket, modifiedObject)) {
+                Matcher matcher = Pattern.compile(".*/(\\d{4})/(nodelist\\.\\d{3})").matcher(modifiedObject);
+                if (!matcher.find()) {
+                    log.error("Invalid nodelist path format: {}", modifiedObject);
+                    return;
+                }
 
-            log.info("Starting updating nodelist {} in database", modifiedObject);
-            nodeEntryRepository.deleteAll();
-            nodelistEntryRepository.deleteAll();
-            updateNodelist(nodelist, Integer.parseInt(matcher.group(1)), matcher.group(2));
-            log.info("Nodelist {} is added to database", modifiedObject);
+                Nodelist nodelist = new Nodelist(new ByteArrayInputStream(inputStream.readAllBytes()));
 
-            log.info("Clearing redis cache");
-            clearRedisCache.clearCache();
-            log.info("Redis cache is cleared");
-        } catch (Exception e) {
-            log.error("Failed to add nodelist to database", e);
+                log.info("Starting updating nodelist {} in database", modifiedObject);
+                nodeEntryRepository.deleteAll();
+                updateNodelist(nodelist, Integer.parseInt(matcher.group(1)), matcher.group(2));
+                log.info("Nodelist {} is added to database", modifiedObject);
+
+                log.info("Clearing redis cache");
+                clearRedisCache.clearCache();
+                log.info("Redis cache is cleared");
+            } catch (Exception e) {
+                log.error("Failed to add nodelist to database", e);
+            }
+        } catch (NoNewObjects e) {
+            log.warn("No new nodelist objects found");
         }
-
-        log.info("Update nodelists is finished");
     }
 
     /**
@@ -117,18 +121,11 @@ public class NodelistFillToDatabase {
      */
     @Transactional
     private void updateNodelist(Nodelist nodelist, Integer year, String name) {
-        if (!nodelistEntryRepository.existsByNodelistYearAndNodelistName(year, name)) {
-            log.info("Update nodelist from {} year and name \"{}\" is started", year, name);
+        log.info("Update nodelist from {} year and name \"{}\" is started", year, name);
 
-            NodelistEntry nodelistEntryNew = new NodelistEntry();
-            nodelistEntryNew.setNodelistYear(year);
-            nodelistEntryNew.setNodelistName(name);
-            nodelistEntryRepository.save(nodelistEntryNew);
-
-            for (ru.oldzoomer.nodelistj.entries.NodelistEntry nodeListEntry : nodelist.getNodelist()) {
-                nodeEntryRepository.save(getNodeEntry(nodeListEntry, nodelistEntryNew));
-            }
-            log.info("Update nodelist from {} year and name \"{}\" is finished", year, name);
+        for (ru.oldzoomer.nodelistj.entries.NodelistEntry nodeListEntry : nodelist.getNodelist()) {
+            nodeEntryRepository.save(getNodeEntry(nodeListEntry, year, name));
         }
+        log.info("Update nodelist from {} year and name \"{}\" is finished", year, name);
     }
 }
